Turn KernelFactory into mixin class for parameter selection
This is potentially only a temporary solution until we find a better
design or figure out additional requirements. The class is private
to simplify future refactorings.
