+++ baybe/surrogates/gaussian_process/components/__init__.py
@@ -1,7 +1,6 @@
 """Gaussian process surrogate components."""
 
 from baybe.surrogates.gaussian_process.components.kernel import (
-    KernelFactory,
     KernelFactoryProtocol,
     PlainKernelFactory,
 )
@@ -17,7 +16,6 @@ from baybe.surrogates.gaussian_process.components.mean import (
 
 __all__ = [
     # Kernel
-    "KernelFactory",
     "KernelFactoryProtocol",
     "PlainKernelFactory",
     # Likelihood
